[PRTL-2702] abstract shared utilities (#2640)
