+++ src/packages/@ncigdc/modern_components/ClinicalAnalysis/ClinicalVariableCard/modals/ContinuousCustomBinsModal/ContinuousCustomBinsModal.js
@@ -1,11 +1,21 @@
 import React, { Component } from 'react';
-import { debounce, isEmpty, isFinite } from 'lodash';
+import {
+  debounce, isEqual, isFinite,
+} from 'lodash';
+import Button from '@ncigdc/uikit/Button';
+import Undo from '@ncigdc/theme/icons/Undo';
+
 
 import { Row, Column } from '@ncigdc/uikit/Flex';
-import Button from '@ncigdc/uikit/Button';
-import { createContinuousGroupName } from '@ncigdc/utils/string';
-import Undo from '@ncigdc/theme/icons/Undo';
+
 import { theme } from '@ncigdc/theme/index';
+import {
+  createContinuousGroupName,
+  DEFAULT_BIN_TYPE,
+  DEFAULT_DATA,
+  DEFAULT_INTERVAL,
+  DEFAULT_RANGES,
+} from '../../helpers';
 import RangeTableRow from './RangeTableRow';
 import BinningMethodInput from './BinningMethodInput';
 import CustomIntervalFields from './CustomIntervalFields';
@@ -32,24 +42,18 @@ const rangeFieldsOrder = [
   'to',
 ];
 
-const defaultInterval = {
-  amount: '',
-  max: '',
-  min: '',
-};
-
 const defaultState = {
   binningMethod: 'interval', // interval or range
   continuousReset: false,
-  intervalErrors: defaultInterval,
-  intervalFields: defaultInterval,
+  intervalErrors: DEFAULT_INTERVAL,
+  intervalFields: DEFAULT_INTERVAL,
   modalWarning: '',
   rangeInputErrors: defaultRangeFieldsState,
   rangeInputOverlapErrors: [],
   rangeInputValues: defaultRangeFieldsState,
   rangeNameErrors: [],
   rangeOverlapErrors: [],
-  rangeRows: [],
+  rangeRows: DEFAULT_RANGES,
 };
 
 class ContinuousCustomBinsModal extends Component {
@@ -57,10 +61,10 @@ class ContinuousCustomBinsModal extends Component {
 
   componentDidMount = () => {
     const {
-      continuousBinType = 'default',
-      continuousCustomInterval = defaultInterval,
-      continuousCustomRanges = [],
-      defaultContinuousData,
+      continuousBinType = DEFAULT_BIN_TYPE,
+      customInterval = DEFAULT_INTERVAL,
+      customRanges = DEFAULT_RANGES,
+      defaultData = DEFAULT_DATA,
     } = this.props;
 
     this.debounceValidateIntervalFields = debounce(
@@ -72,26 +76,26 @@ class ContinuousCustomBinsModal extends Component {
       ...continuousBinType === 'default'
         ? {}
         : { binningMethod: continuousBinType },
-      intervalFields: isEmpty(continuousCustomInterval)
+      intervalFields: isEqual(customInterval, DEFAULT_INTERVAL)
         ? {
-          amount: defaultContinuousData.quarter,
-          max: defaultContinuousData.max,
-          min: defaultContinuousData.min,
+          amount: defaultData.quarter,
+          max: defaultData.max,
+          min: defaultData.min,
         }
-        : continuousCustomInterval,
-      rangeRows: continuousCustomRanges,
+        : customInterval,
+      rangeRows: customRanges,
     });
   };
 
   resetModal = () => {
-    const { defaultContinuousData } = this.props;
+    const { defaultData } = this.props;
     this.setState({
       ...defaultState,
       continuousReset: true,
       intervalFields: {
-        amount: defaultContinuousData.quarter,
-        max: defaultContinuousData.max,
-        min: defaultContinuousData.min,
+        amount: defaultData.quarter,
+        max: defaultData.max,
+        min: defaultData.min,
       },
     });
   }
@@ -335,12 +339,15 @@ class ContinuousCustomBinsModal extends Component {
           const rowKey = `${curr.from}-${curr.to}`;
           return ({
             ...acc,
-            ...(curr.name === '' ? {} : {
-              [rowKey]: {
-                groupName: curr.name,
-                key: rowKey,
-              },
-            }),
+            ...(curr.name === ''
+              ? {}
+              : {
+                [rowKey]: {
+                  groupName: curr.name,
+                  key: rowKey,
+                },
+              }
+            ),
           });
         }, {})
         : makeCustomIntervalBins();
@@ -573,7 +580,7 @@ class ContinuousCustomBinsModal extends Component {
 
 
   render = () => {
-    const { defaultContinuousData, fieldName, onClose } = this.props;
+    const { defaultData, fieldName, onClose } = this.props;
     const {
       binningMethod,
       intervalErrors,
@@ -609,14 +616,14 @@ class ContinuousCustomBinsModal extends Component {
           <Row style={styles.defaultInfo}>
             <p style={styles.defaultInfo.paragraph}>
               Available values from
-              <strong>{` ${defaultContinuousData.min} `}</strong>
+              <strong>{` ${defaultData.min} `}</strong>
               to
-              <strong>{` \u003c ${defaultContinuousData.max} `}</strong>
+              <strong>{` \u003c ${defaultData.max} `}</strong>
             </p>
             <p style={styles.defaultInfo.paragraph}>|</p>
             <p style={styles.defaultInfo.paragraph}>
               Bin size in quarters:
-              <strong>{` ${defaultContinuousData.quarter}`}</strong>
+              <strong>{` ${defaultData.quarter}`}</strong>
             </p>
           </Row>
         </div>
