+++ src/packages/@ncigdc/modern_components/ClinicalAnalysis/ContinuousAggregationQuery.js
@@ -11,11 +11,12 @@ import _ from 'lodash';
 
 import consoleDebug from '@ncigdc/utils/consoleDebug';
 import { redirectToLogin } from '@ncigdc/utils/auth';
-import { createFacetFieldString, parseContinuousKey } from '@ncigdc/utils/string'
+import { createFacetFieldString } from '@ncigdc/utils/string'
 import Loader from '@ncigdc/uikit/Loaders/Loader';
 
 import { API, IS_AUTH_PORTAL } from '@ncigdc/utils/constants';
-import ClinicalVariableCard from './ClinicalVariableCard';
+import { ContinuousVariableCard } from './ClinicalVariableCard';
+import { parseContinuousKey } from './ClinicalVariableCard/helpers';
 
 const simpleAggCache = {};
 const pendingAggCache = {};
@@ -41,6 +42,7 @@ const getContinuousAggs = ({
         from: key * interval + stats.min,
         to: (key + 1) === DEFAULT_CONTINUOUS_BUCKETS
           ? stats.max + 1
+          // api excludes max value
           : stats.min + (key + 1) * interval,
       })
     );
@@ -222,7 +224,7 @@ export default compose(
   }
 
   return (
-    <ClinicalVariableCard
+    <ContinuousVariableCard
       data={{
         ...aggData,
         hits,
