+++ src/packages/@ncigdc/modern_components/ClinicalAnalysis/survivalPlot.css
@@ -1,29 +1,29 @@
 /* clinical analysis */
 
-.clinical-analysis-categorical-card .test-visualization-title[style] {
+.clinical-analysis-card .test-visualization-title[style] {
   display: none !important;
 }
 
-.clinical-analysis-categorical-card .test-visualization-title + div[style] {
+.clinical-analysis-card .test-visualization-title + div[style] {
   display: block !important;
 }
 
-.clinical-analysis-categorical-card .test-visualization-header[style] {
+.clinical-analysis-card .test-visualization-header[style] {
   display: block !important;
   margin: 0 !important;
   text-align: right;
   width: 100%;
 }
 
-.clinical-analysis-categorical-card .test-survival-plot-meta[style] {
+.clinical-analysis-card .test-survival-plot-meta[style] {
   display: block !important;
 }
 
-.clinical-analysis-categorical-card .p-value {
+.clinical-analysis-card .p-value {
   margin-right: 40px;
 }
 
-.clinical-analysis-categorical-card td:nth-child(2)[style] {
+.clinical-analysis-card td:nth-child(2)[style] {
   /* wrap long facet values so the table doesn't scroll */
   white-space: normal !important;
 }
